pagination and favorites
